# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce API
  version: 1.0.0
  description: API First for ecommerce project
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/auth/register-admin:
    post:
      tags:
        - Auth
      operationId: registerAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterApiRequest"
      responses:
        "201":
          description: Created
...
  /api/products:
     get:
        tags:
          - Product
        operationId: getAllProducts
        security: []
        parameters:
          - name: page
            in: query
            required: false
            schema:
              type: integer
              example: 0
          - name: size
            in: query
            required: false
            schema:
              type: integer
              example: 10
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/PageProductApiResponse"
  /api/products/{id}:
      get:
        tags:
          - Product
        operationId: getProductById
        security: []
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ProductApiResponse"
  /api/products/category/{categoryId}:
    get:
        tags:
          - Product
        operationId: getProductsByCategory
        security: []
        parameters:
          - name: categoryId
            in: path
            required: true
            schema:
              type: integer
          - name: page
            in: query
            required: false
            schema:
              type: integer
          - name: size
            in: query
            required: false
            schema:
              type: integer
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/PageProductApiResponse"
  /api/admin/products:
    post:
        tags:
          - Admin
        operationId: createProduct
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiRequest"
        responses:
          "201":
            description: Created
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ProductApiResponse"
  /api/admin/products/{id}:
    put:
        tags:
          - Admin
        operationId: updateProduct
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiRequest"
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ProductApiResponse"

    delete:
        tags:
          - Admin
        operationId: deleteProduct
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        responses:
          "204":
            description: Deleted
...
  /api/admin/categories:
      post:
        tags:
          - Admin
        operationId: createCategory
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiRequest"
        responses:
          "201":
            description: Created
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CategoryApiResponse"
  /api/admin/categories/{id}:
      put:
        tags:
          - Admin
        operationId: updateCategory
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiRequest"
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CategoryApiResponse"
...
  /api/orders/{id}/items:
    post:
        tags:
          - Order
        operationId: addItemToOrder
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItemApiRequest"
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/OrderApiResponse"
...
  /api/orders/{id}:
      get:
        tags:
          - Order
        operationId: getOrderById
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/OrderApiResponse"
...
  /api/admin/orders/{id}/status:
    put:
        tags:
          - Admin
        operationId: updateOrderStatus
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/OrderApiResponse"

    delete:
        tags:
          - Admin
        operationId: deleteCategory
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        responses:
          "204":
            description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterApiRequest:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: string
...
    ProductApiResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        categoryId:
          type: integer
    ProductApiRequest:
      type: object
      required:
        - name
        - price
        - categoryId
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        categoryId:
          type: integer
    PageProductApiResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductApiResponse"
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    CategoryApiRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    CategoryApiResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
    OrderApiResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        totalPrice:
          type: number
          format: double
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemApiResponse"
    OrderItemApiRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    OrderItemApiResponse:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: double